feat(eds-color-palette-generator): ✨ refactor
